requirements-precommit.txt
requirements-parser==0.5.0
setuptools==44.0.0 # Must be fixed to this as later versions have a bug on reqs that include "+cpu"
